dashboard/collector/requirements.txt
psycopg2-binary==2.9.9
PyYAML==6.0.1
